requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
stripe
python-jose
passlib[bcrypt]
psycopg2-binary
